quick updates in structure
